webkit/data/layout_tests/chrome/http/tests/security/listener/xss-inactive-closure-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/listener/resources/xss-inactive-closure-child-2.html from frame with URL http://127.0.0.1:8081/chrome/http/tests/security/listener/resources/childWithButton.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 6: Uncaught TypeError: Cannot read property 'bar' of undefined
This tests that when a frame navigates to a new page, closures in the old page cannot access page content of the new page if there are from different domains.

You should see PASSSED at the end of page if the test passes.

  
PASSED
